restructure test files
